[refactor] dropdown -> select, 재사용성 높이기
