--- src/common/responsive.js
@@ -1,19 +1,0 @@
-import React from "react";
-import { useMediaQuery } from 'react-responsive';
- 
-export const Desktop = ({ children }) => {
-  const isDesktop = useMediaQuery({ minWidth: 1200 })
-  return isDesktop ? children : null
-}
-export const Tablet = ({ children }) => {
-  const isTablet = useMediaQuery({ minWidth: 768, maxWidth: 1119 })
-  return isTablet ? children : null
-}
-export const Mobile = ({ children }) => {
-  const isMobile = useMediaQuery({ maxWidth: 767 })
-  return isMobile ? children : null
-}
-export const Default = ({ children }) => {
-  const isNotMobile = useMediaQuery({ minWidth: 768 })
-  return isNotMobile ? children : null
-}
